Upgrade to v5 (#18)
Update the base github provider to ~> v5.0

The providers major version change is related to how branches are specified in this version. We dont normally specifiy non-default branches in terraform, so will be a low impact on our code. However, as its a major version change in the provider, i've done the same for our version as well.

This change also includes removal of a lifecycle hack where we ignored changes on list of branches - as this property no longer exists on the repository resource.

#major